+++ library/lib/core/application.cpp
@@ -208,7 +208,7 @@ bool Application::mainLoop()
     }
 
     std::vector<TouchState> touchState;
-    for (int i = 0; i < rawTouch.size(); i++)
+    for (size_t i = 0; i < rawTouch.size(); i++)
     {
         auto old = std::find_if(std::begin(currentTouchState), std::end(currentTouchState), [rawTouch, i](TouchState touch)
             { return touch.fingerId == rawTouch[i].fingerId; });
@@ -225,7 +225,7 @@ bool Application::mainLoop()
         }
     }
 
-    for (int i = 0; i < currentTouchState.size(); i++)
+    for (size_t i = 0; i < currentTouchState.size(); i++)
     {
         if (currentTouchState[i].phase == TouchPhase::NONE)
             continue;
@@ -239,7 +239,7 @@ bool Application::mainLoop()
         }
     }
 
-    for (int i = 0; i < touchState.size(); i++)
+    for (size_t i = 0; i < touchState.size(); i++)
     {
         if (touchState[i].phase == TouchPhase::NONE)
         {
@@ -872,7 +872,7 @@ void Application::tryDeinitFirstResponder(View* view)
         return;
 
     // Interrupt current gestures if presented
-    for (int i = 0; i < currentTouchState.size(); i++)
+    for (size_t i = 0; i < currentTouchState.size(); i++)
     {
         if (currentTouchState[i].view == view)
         {
